% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_multi_drop.R
\name{add_multi_drop}
\alias{add_multi_drop}
\title{add multiple dropdowns automatically to existing highcharter plots}
\usage{
add_multi_drop(hc, selectors, selected = NULL)
}
\arguments{
\item{hc}{highcharts object}

\item{selectors}{the keys which to use for select dropdowns}

\item{selected}{optional selected keys}
}
\value{
returns a browsable html object
}
\description{
add multiple dropdowns automatically to existing highcharter plots
}
\seealso{
https://jkunst.com/highcharter/
}
