% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_id.R
\name{random_id}
\alias{random_id}
\title{random id}
\usage{
random_id(n = 1, length = 10)
}
\arguments{
\item{n}{number of ids}

\item{length}{length of id}
}
\description{
random id
}
\examples{
random_id(n = 1, length = 10)
}
\seealso{
pulled from highcharter non exports
}
